refactor: move centralized training to own file
